requirements/package.txt
nltk==3.2.5
Flask==0.12.2
redis==2.10.6
python-dotenv==0.7.1
